--- test/snapshots/index.js.md
@@ -1,27 +1,0 @@
-# Snapshot report for `test/index.js`
-
-The actual snapshot is saved in `index.js.snap`.
-
-Generated by [AVA](https://avajs.dev).
-
-## returns html for markdown
-
-> Snapshot 1
-
-    `<h1>Beep</h1>␊
-    <p><strong>Boop!</strong> from <a href="https://hackclub.com/" class="external">Hack Club</a>.</p>`
-
-## returns html code sh markdown
-
-> Snapshot 1
-
-    `<pre class="language-shell"><code class="language-shell">const hi <span class="token operator">=</span> <span class="token punctuation">(</span><span class="token punctuation">)</span> <span class="token operator">=</span><span class="token operator">></span> console.log<span class="token punctuation">(</span><span class="token string">'Hello!'</span><span class="token punctuation">)</span>␊
-    </code></pre>`
-
-## returns html code for video link
-
-> Snapshot 1
-
-    `<p>This should not be rendered: https://user-images.githubusercontent.com/27078897/141502788-e4687f86-2d43-458b-a524-c420fba09ffd.mov</p>␊
-    <p>But this should!</p>␊
-    <details open class="details-video"><summary><div class="details-video-summary"><span class="details-video-caret"></span><svg aria-hidden height="16" width="16" viewBox="0 0 16 16" version="1.1" class="video-summary-camera-icon"><path fill-rule="evenodd" d="M16 3.75a.75.75 0 00-1.136-.643L11 5.425V4.75A1.75 1.75 0 009.25 3h-7.5A1.75 1.75 0 000 4.75v6.5C0 12.216.784 13 1.75 13h7.5A1.75 1.75 0 0011 11.25v-.675l3.864 2.318A.75.75 0 0016 12.25v-8.5zm-5 5.075l3.5 2.1v-5.85l-3.5 2.1v1.65zM9.5 6.75v-2a.25.25 0 00-.25-.25h-7.5a.25.25 0 00-.25.25v6.5c0 .138.112.25.25.25h7.5a.25.25 0 00.25-.25v-4.5z"></path></svg><span aria-label="141502788-e4687f86-2d43-458b-a524-c420fba09ffd.mov">141502788-e4687f86-2d43-458b-a524-c420fba09ffd.mov</span></div></summary><video src="https://user-images.githubusercontent.com/27078897/141502788-e4687f86-2d43-458b-a524-c420fba09ffd.mov" muted controls style="max-height:640px;"></video></details>`
